changed mail validation
