+++ front-end/src/component/page/register/register.component.css
@@ -1,7 +1,7 @@
 
 .register-form-container {
   width: 500px;
-  margin-left: 65%;
+  margin: 0 auto;
   margin-top: 64px;
 }
