+++ front-end/src/validate/validator.module.ts
@@ -1,8 +1,7 @@
 import {NgModule} from '@angular/core';
 import {CommonModule} from '@angular/common';
-import {EqualsValidatorDirective} from "./equalsValidatorDerective";
-
-
+import {EqualsValidatorDirective} from "./equals-validator.derective";
+import {EmailValidatorDirective} from "./email-validator.directive";
 
 @NgModule({
   imports: [
@@ -10,10 +9,12 @@ import {EqualsValidatorDirective} from "./equalsValidatorDerective";
   ],
   declarations: [
     EqualsValidatorDirective,
+    EmailValidatorDirective
 
   ],
   exports: [
     EqualsValidatorDirective,
+    EmailValidatorDirective,
     CommonModule
   ]
 })
